(refactor) Move ApplicationFixture and ApplicationAdapter to testfx-junit.
