+++ subprojects/testfx-junit/src/main/java/org/testfx/framework/junit/ApplicationFixture.java
@@ -14,7 +14,7 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the Licence for the
  * specific language governing permissions and limitations under the Licence.
  */
-package org.testfx.toolkit;
+package org.testfx.framework.junit;
 
 import javafx.stage.Stage;
